Spike Re-Compose Mapbox Elements on Page
This spike explores how we might re-compose the arrangement of the control panel and other elements for Mapbox over the mapbox page. It also does some initial updates to the map.scss style to override the default mapbox gl style.

In this update the control panel and geolocate components can sit in the same DOM tree area. However the geocoder component plays in a different part of the DOM for unknown reasons.